PDTY-5689 Fix mangled type references (#16)
There are times when a type reference is printed prefixed with the
declaration path and a dollar sign, so instead of:

```ts
type Foo {
  bar: Baz
}
```

we get:

```ts
type Foo {
  bar: "path/to/baz"$Baz
}
```

The patch here feels like a brute force way of fixing the issue. There's
probably something upstream in terms of how things are being transformed
in the AST which could better fix this, but this appears to work.