PDTY-5722: Support inline `type` imports
This change updates the `import` printing logic to properly handle inline
`type` imports, in addition to the already handled top-level `type` imports.

Prior to this change, the behavior was to attempt to import the `type` keyword
itself:

```
import { type, Foo, Bar } from 'place';
```

With this change, we get the more correct:

```
import { Bar } from 'place';
import type { Foo } from 'place';
```

I believe the issue here may be related to the version of the TypeScript parser
being used in this repo. Upgrading the version of TypeScript is something we
could consider, but may have larger rammifications in this codebase.