PDTY-5726: Special-case FlowLang.ObjMap (#21)
When attempting to convert `packages/utilities/css-properties-type`,
there is an instance of `FlowLang.ObjMap` in the generated interface.
This is a type generated by the `flow-to-typescript-codemod` to attempt
to match `flow`s builtin `$ObjMap` type. As such, we need to special
case this type when generating `.js.flow` interface files to return an
`$ObjMap` as expected.