PDTY-6238: use  for certain Indexed Access Types (#26)
In order to support the upgrade to TypeScript 4.9.5 in the monorepo, we
want to convert the following:

```
(typeof X)[keyof typeof X]
```
into
```
$Values<X>
```
or
```
$Values<typeof X>
```
depending on whether `X` is a type or a value.

without this, we get `$ElementType<X, $Keys<X>>`, which looks like it
should work, but in the version of flow we use seems to be an
intersection, rather than a union, meaning that it is completely
unrepresentable if `X` has more than a single key.

I believe this change should be completely safe to ship independently of
the typescript upgrade. The problem it prevents only seems to manifest
after the upgrade, but the solution should be valid regardless.